requirements.txt
torch
torchvision
pillow
wandb
matplotlib